Реализовал модуль timer_util
- Реализовал функцию timer_get_tick

+++ Dual_channel_LPS/inc/timer/timer_util.h
@@ -9,6 +9,6 @@
 #ifndef TIMER_UTIL_H_
 #define TIMER_UTIL_H_
 
-unsigned int timer_get_tick();
+unsigned long timer_get_tick();
 
 #endif /* TIMER_UTIL_H_ */
